+++ .gitattributes
@@ -1,1 +1,1 @@
-src/scoping/langs/tree_sitter_hcl/upstream-main/** linguist-vendored
+src/langs/tree_sitter_hcl/upstream-main/** linguist-vendored
