+++ NOTICE
@@ -1,3 +1,3 @@
 This project includes source code developed by the [Tree-sitter Grammars]
 (https://github.com/tree-sitter-grammars) organization. See
-[tree-sitter-hcl](src/scoping/langs/tree_sitter_hcl/README.md) for more.
+[tree-sitter-hcl](src/langs/tree_sitter_hcl/README.md) for more.
