+++ src/langs/hcl.rs
@@ -5,7 +5,7 @@ use const_format::formatcp;
 
 use super::{tree_sitter_hcl, LanguageScoper, RawQuery, TSLanguage, TSQuery, TSQueryError};
 use crate::find::Find;
-use crate::scoping::langs::IGNORE;
+use crate::langs::IGNORE;
 
 /// A compiled query for the Hashicorp Configuration language.
 #[derive(Debug)]
