+++ src/langs/tree_sitter_hcl/upstream-main/binding.gyp
