+++ src/langs/tree_sitter_hcl/upstream-main/bindings/node/index.js
