+++ src/langs/tree_sitter_hcl/upstream-main/dialects/terraform/src/node-types.json
